ADPIRA

Analysis Dataset for the study of Progression Independent of Relapse Activity in the Polish RRMS patient population.

ADPIRA

Format

Information about the fornat

Source

UMB.

Examples

attr(ADPIRA, "metadata")

#> NULL

sapply(ADPIRA, attr, which = "label")

#> $ARM
#> [1] "Arm of the study, the prescription panel."
#> 
#> $ARMCD
#> [1] "Short code for the study Arm."
#> 
#> $EDSS_start
#> [1] "EDSS at prescription start."
#> 
#> $treatment_line
#> [1] "Treatment line."
#> 
#> $prescription
#> [1] "Drug prescribed, one of Interferon beta (INF), Glatiramer Acetate (GA), Fingolimod (FTY), Natalizumab (NAT)."
#> 
#> $presc_months_cat
#> [1] "Monitoring time point in months since prescription start (12, 24, 36, 48, 60 months)"
#> 
#> $EDSS_current
#> [1] "EDSS at the time of the visit."
#> 
#> $dMRI_12m
#> [1] "Brain MRI activity: number of new T2 or Gd+ lesion detected within 12 months preceeding the visit."
#> 
#> $worse
#> [1] "The disease was worse than at baseline (TRUE/FALSE)."
#> 
#> $sex
#> [1] "Sex."
#> 
#> $age
#> NULL
#> 
#> $AgeSympt
#> [1] "Birthyear"
#> 
#> $Sympt2Presc
#> NULL
#> 

head(ADPIRA)

#>               ARM ARMCD EDSS_start treatment_line prescription presc_months_cat
#> 1  R0: no relapse    R0        0.5              I          INF               12
#> 10 R0: no relapse    R0        3.0              I          INF               12
#> 17 R0: no relapse    R0        0.0              I          GLA               12
#> 22 R0: no relapse    R0        1.0              I          INF               12
#> 24 R0: no relapse    R0        2.0              I          INF               12
#> 26 R0: no relapse    R0        2.5              I          INF               12
#>    EDSS_current dMRI_12m worse    sex age AgeSympt Sympt2Presc
#> 1           0.5        3 FALSE   male  24       24         200
#> 10          3.0        0 FALSE female  54       52         986
#> 17          0.0        0 FALSE female  25       24         543
#> 22          1.0        1 FALSE   male  25       23         919
#> 24          1.5        0 FALSE   male  37       33        1627
#> 26          2.0        0 FALSE female  52       43        3124